requirements.txt
fastapi
uvicorn[standard]
numpy
Pillow
opencv-python-headless
python-multipart
reportlab
Jinja2
